+++ ach/src/test/java/com/adyen/checkout/ach/internal/ui/DefaultACHDirectDebitDelegateTest.kt
@@ -606,6 +606,15 @@ internal class DefaultACHDirectDebitDelegateTest(
     }
 
     @Nested
+    inner class SubmitButtonEnableTest {
+
+        @Test
+        fun `when shouldEnableSubmitButton is called, then true is returned`() {
+            assertTrue(delegate.shouldEnableSubmitButton())
+        }
+    }
+
+    @Nested
     inner class SubmitHandlerTest {
 
         @Test