+++ boleto/src/main/java/com/adyen/checkout/boleto/internal/ui/DefaultBoletoDelegate.kt
@@ -303,6 +303,8 @@ internal class DefaultBoletoDelegate(
 
     override fun shouldShowSubmitButton(): Boolean = isConfirmationRequired() && componentParams.isSubmitButtonVisible
 
+    override fun shouldEnableSubmitButton(): Boolean = true
+
     override fun onCleared() {
         removeObserver()
         analyticsManager.clear(this)