+++ sepa/src/test/java/com/adyen/checkout/sepa/internal/ui/DefaultSepaDelegateTest.kt
@@ -151,6 +151,15 @@ internal class DefaultSepaDelegateTest(
     }
 
     @Nested
+    inner class SubmitButtonEnableTest {
+
+        @Test
+        fun `when shouldEnableSubmitButton is called, then true is returned`() {
+            assertTrue(delegate.shouldEnableSubmitButton())
+        }
+    }
+
+    @Nested
     inner class SubmitHandlerTest {
 
         @Test