+++ ui-core/src/main/java/com/adyen/checkout/ui/core/internal/test/TestRedirectHandler.kt
@@ -40,6 +40,9 @@ class TestRedirectHandler : RedirectHandler {
     fun assertLaunchRedirectCalled() =
         assert(timesLaunchRedirectCalled > 0)
 
+    fun assertLaunchRedirectNotCalled() =
+        assert(timesLaunchRedirectCalled <= 0)
+
     override fun setOnRedirectListener(listener: () -> Unit) = Unit
 
     override fun removeOnRedirectListener() {